backend/pyproject.toml
[tool.poetry]
name = "ansible-ui-backend"
version = "0.2.0"
description = "Django REST Framework backend for Ansible UI - processes Ansible logs and provides REST API"
authors = ["Ansible UI Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14"
django-cors-headers = "^4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
black = "^23.12"
flake8 = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
